:construction: Get rsbuild to serve page
