--- jest-preprocess.js
@@ -1,5 +1,0 @@
-const babelOptions = {
-  presets: ['babel-preset-gatsby', '@babel/preset-typescript'],
-};
-
-module.exports = require('babel-jest').default.createTransformer(babelOptions);
