--- jest.config.js
@@ -1,25 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.[jt]sx?$': '<rootDir>/jest-preprocess.js',
-  },
-  moduleNameMapper: {
-    '.+\\.(css|styl|less|sass|scss)$': `identity-obj-proxy`,
-    '.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': `<rootDir>/__mocks__/file-mock.js`,
-  },
-  testPathIgnorePatterns: [`node_modules`, `\\.cache`, `<rootDir>.*/public`],
-  transformIgnorePatterns: [
-    `node_modules/(?!(gatsby|gatsby-script|gatsby-link)/)`,
-  ],
-  globals: {
-    __PATH_PREFIX__: ``,
-  },
-  testEnvironment: 'jsdom',
-  testEnvironmentOptions: {
-    url: `http://localhost`,
-  },
-  setupFiles: [`<rootDir>/loadershim.js`],
-  collectCoverage: true,
-  collectCoverageFrom: ['src/**/*.{js,tsx}'],
-  coveragePathIgnorePatterns: ['<rootDir>/node_modules/'],
-  setupFilesAfterEnv: ['<rootDir>/setupTests.ts'],
-};
