+++ source.entry
@@ -1,0 +1,3 @@
+export default {
+  index: './src/index.tsx'
+};
