+++ src/components/Footer/index.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import * as styles from './Footer.module.css';
+import styles from './Footer.module.css';
 
 export default function Footer() {
   return (
