+++ src/components/SEO.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function Seo({
   description,
   imageUrl,
